# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Makers Community API",
        "description": "API complète pour la plateforme communautaire Makers Community avec authentification, posts, commentaires, ratings et modération automatique. Cette API permet de gérer une communauté de makers avec toutes les fonctionnalités nécessaires.",
        "contact": {
            "name": "Équipe Makers Community",
            "email": "dev@makerscomm.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Serveur de développement local"
        },
        {
            "url": "https://api.makerscomm.com/api",
            "description": "Serveur de production"
        }
    ],
    "paths": {
...
        "/posts": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Lister tous les posts",
                "description": "Récupérer tous les posts de la communauté avec les informations utilisateur",
                "operationId": "5f292ed47bdbfa79356750be0807450c",
                "responses": {
                    "200": {
                        "description": "Liste des posts récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "error": {
                                            "type": "null"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Post"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "List of posts retrieved"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Créer un nouveau post",
                "description": "Créer un nouveau post dans la communauté",
                "operationId": "c2e0b0e26215e6d34ece33ebae25f7df",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "title",
                                    "content"
                                ],
                                "properties": {
                                    "user_id": {
                                        "description": "ID de l'utilisateur créateur",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "title": {
                                        "description": "Titre du post",
                                        "type": "string",
                                        "example": "Mon projet Arduino"
                                    },
                                    "content": {
                                        "description": "Contenu du post",
                                        "type": "string",
                                        "example": "Description détaillée de mon projet..."
                                    },
                                    "link": {
                                        "description": "Lien externe optionnel",
                                        "type": "string",
                                        "example": "https://github.com/user/projet",
                                        "nullable": true
                                    },
                                    "photos": {
                                        "description": "Liste des photos (noms séparés par virgules)",
                                        "type": "string",
                                        "example": "image1.jpg,image2.jpg",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Post créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 201
                                        },
                                        "error": {
                                            "type": "null"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Post"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Post created successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "User": {
                "title": "Utilisateur",
                "description": "Modèle utilisateur de base",
                "properties": {
                    "id": {
                        "description": "Identifiant unique",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Nom complet",
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "description": "Adresse email",
                        "type": "string",
                        "format": "email",
                        "example": "john@example.com"
                    },
                    "birthday": {
                        "description": "Date de naissance",
                        "type": "string",
                        "format": "date",
                        "example": "1995-05-15"
                    },
                    "role": {
                        "description": "Rôle utilisateur",
                        "type": "string",
                        "enum": [
                            "user",
                            "admin"
                        ],
                        "example": "user"
                    },
                    "profile_pic": {
                        "description": "Photo de profil",
                        "type": "string",
                        "example": "profile_123.jpg",
                        "nullable": true
                    },
                    "is_active": {
                        "description": "Statut actif du compte",
                        "type": "boolean",
                        "example": true
                    },
                    "email_verified_at": {
                        "description": "Date de vérification email",
                        "type": "string",
                        "format": "datetime",
                        "example": "2023-10-02T14:30:00Z",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "datetime",
                        "example": "2023-10-01T10:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "datetime",
                        "example": "2023-10-02T14:30:00Z"
                    }
                },
                "type": "object"
            },
...
            "Post": {
                "title": "Post",
                "description": "Publication communautaire",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Mon premier projet Arduino"
                    },
                    "content": {
                        "type": "string",
                        "example": "Voici la description détaillée de mon projet..."
                    },
                    "link": {
                        "type": "string",
                        "example": "https://github.com/user/projet",
                        "nullable": true
                    },
                    "photos": {
                        "type": "string",
                        "example": "post_123.jpg,post_124.jpg",
                        "nullable": true
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "datetime",
                        "example": "2023-10-01T10:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "datetime",
                        "example": "2023-10-02T14:30:00Z"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object"
            },
...
            "ValidationError": {
                "title": "Erreur de Validation",
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "error"
                    },
                    "error": {
                        "properties": {
                            "field_name": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "example": "Ce champ est requis."
                                }
                            }
                        },
                        "type": "object"
                    },
                    "data": {
                        "type": "null"
                    },
                    "message": {
                        "type": "string",
                        "example": "Erreur de validation"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token d'authentification Laravel Sanctum. Format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}